requirements.txt
streamlit==1.39.0
pandas>=2.0
gspread==6.1.4
google-auth==2.35.0
